Moved the tests.
