--- .vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "restructuredtext.confPath": "${workspaceFolder}/doc",
-    "restructuredtext.linter.disabled": true
-}
