+++ geomloss/ot/__init__.py
@@ -1,5 +1,3 @@
 from .implementations.matrix import solve, barycenter
 from .implementations.samples import solve_samples, barycenter_samples
 from .implementations.grid import solve_grid, barycenter_grid
-
-from . import tests
