+++ tests/generators/diracs.py
