+++ tests/generators/permutations.py
